requirements.txt
aiogram==2.10.1
python-dotenv
SpeechRecognition
ftransc